KB-5545 | DEV| Assessment | BE | Enhancement in Consumption Logic for the OptionWeightage Assessment Type.
1. casting exception handled.

+++ src/main/java/org/sunbird/assessment/service/AssessmentUtilServiceV2Impl.java
@@ -686,7 +686,7 @@ public class AssessmentUtilServiceV2Impl implements AssessmentUtilServiceV2 {
 	private Double handleIncorrectAnswer(int negativeMarksValue,Double sectionMarks, Map<String, Object> questionSetSectionScheme, Map<String, Object> proficiencyMap) {
 		logger.info("Handling incorrect answer scenario...");
 		if (negativeMarksValue > 0) {
-			sectionMarks = sectionMarks - (((double)negativeMarksValue /100 ) * (double) questionSetSectionScheme.get((String)proficiencyMap.get(Constants.QUESTION_LEVEL)));
+			sectionMarks = sectionMarks - (((double)negativeMarksValue /100 ) * (int) questionSetSectionScheme.get((String)proficiencyMap.get(Constants.QUESTION_LEVEL)));
 		}
 		logger.info("Incorrect answer scenario handled successfully.");
 		return sectionMarks;
